+++ src/rule-definitions.ts
@@ -1,5 +1,7 @@
 import type { RuleDefinition } from '#src/common/rule.js';
 import banUnboundedStringType from '#src/rules/ban-unbounded-string-type.js';
+import enumNamePascalCase from '#src/rules/enum-name-pascal-case.js';
+import enumValueSnakeCase from '#src/rules/enum-value-snake-case.js';
 import fieldNameCamelCase from '#src/rules/field-name-camel-case.js';
 import fieldNameMappingSnakeCase from '#src/rules/field-name-mapping-snake-case.js';
 import fieldOrder from '#src/rules/field-order.js';
@@ -16,6 +18,8 @@ import requireField from '#src/rules/require-field.js';
 
 export default [
   banUnboundedStringType,
+  enumNamePascalCase,
+  enumValueSnakeCase,
   fieldNameCamelCase,
   fieldNameMappingSnakeCase,
   fieldOrder,